refactor: separate concerns cleanly, and bench every OCR engine
